+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to the "fsix-vscode" extension will be documented in this fi
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+- Use tcp sockets instead of stdio. Now stdout in your app doesnt break fsix
+- Dispose of connections and processes for closed notebooks
+- Fixed cancellation support, for both notebooks and repls 
 ## [Unreleased]
 
 - Initial release